+++ src/main/java/cn/memoryzy/json/util/JsonAssistantUtil.java
@@ -1,6 +1,8 @@
 package cn.memoryzy.json.util;
 
+import cn.hutool.core.date.DatePattern;
 import cn.hutool.core.date.DateUtil;
+import cn.hutool.core.date.LocalDateTimeUtil;
 import cn.hutool.core.text.NamingCase;
 import cn.hutool.core.util.*;
 import org.apache.commons.lang3.StringUtils;
@@ -8,6 +10,7 @@ import org.apache.commons.lang3.StringUtils;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
+import java.time.LocalDateTime;
 import java.util.*;
 
 /**
@@ -15,6 +18,12 @@ import java.util.*;
  * @since 2024/8/3
  */
 public class JsonAssistantUtil {
+
+    private static final long MIN_VALID_TIMESTAMP_SECONDS = 0L; // 1970-01-01T00:00:00Z
+    private static final long MAX_VALID_TIMESTAMP_SECONDS = 4102444800L; // 2099-12-31T23:59:59Z
+
+    private static final long MIN_VALID_TIMESTAMP_MILLIS = MIN_VALID_TIMESTAMP_SECONDS * 1000;
+    private static final long MAX_VALID_TIMESTAMP_MILLIS = MAX_VALID_TIMESTAMP_SECONDS * 1000;
 
     public static String truncateText(String text, int maxLength, String omitHint) {
         if (text.length() > maxLength) {
@@ -258,6 +267,65 @@ public class JsonAssistantUtil {
             currentClass = currentClass.getSuperclass();
         }
         return false;
+    }
+
+
+    /**
+     * 根据时间戳的详细信息（时、分、秒、毫秒）选择合适的日期时间格式并返回格式化后的字符串。
+     *
+     * @param timestamp 时间戳（单位：毫秒）
+     * @return 格式化后的日期时间字符串
+     */
+    public static String formatDateBasedOnTimestampDetails(long timestamp) {
+        LocalDateTime localDateTime = LocalDateTimeUtil.of(timestamp);
+
+        int hour = localDateTime.getHour();
+        int minute = localDateTime.getMinute();
+        int second = localDateTime.getSecond();
+        // 获取毫秒部分
+        int millis = (int) ((timestamp % 1000));
+
+        String format;
+        if (hour == 0 && minute == 0 && second == 0 && millis == 0) {
+            // 如果时、分、秒和毫秒都为0，则只显示日期
+            format = DatePattern.NORM_DATE_PATTERN;
+        } else if (millis > 0) {
+            // 如果毫秒部分存在值，则包含毫秒的格式化表达式
+            format = DatePattern.NORM_DATETIME_MS_PATTERN;
+        } else if (second > 0) {
+            // 如果秒部分存在值，则包含秒的格式化表达式
+            format = DatePattern.NORM_DATETIME_PATTERN;
+        } else if (minute > 0 || hour > 0) {
+            // 如果分钟或小时部分存在值，则包含时和分的格式化表达式
+            format = DatePattern.NORM_DATETIME_MINUTE_PATTERN;
+        } else {
+            // 默认情况下仅显示日期
+            format = DatePattern.NORM_DATE_PATTERN;
+        }
+
+        return LocalDateTimeUtil.format(localDateTime, format);
+    }
+
+    /**
+     * 验证是否为合法的时间戳格式
+     *
+     * @param timestampStr 时间戳字符串
+     * @return 合法的时间戳为true，否则为false
+     */
+    public static boolean isValidTimestamp(String timestampStr) {
+        try {
+            long timestamp = Long.parseLong(timestampStr);
+
+            if (timestampStr.length() == 10) { // Seconds-based timestamp
+                return timestamp >= MIN_VALID_TIMESTAMP_SECONDS && timestamp <= MAX_VALID_TIMESTAMP_SECONDS;
+            } else if (timestampStr.length() == 13) { // Milliseconds-based timestamp
+                return timestamp >= MIN_VALID_TIMESTAMP_MILLIS && timestamp <= MAX_VALID_TIMESTAMP_MILLIS;
+            } else {
+                return false; // Invalid length for a timestamp
+            }
+        } catch (NumberFormatException e) {
+            return false; // Not a valid number
+        }
     }
 
 }